backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.3
pillow>=11.0.0
easyocr>=1.7.1
opencv-python>=4.9.0
